pyproject.toml
[project]
name = "skipper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "mlx-audio>=0.4.1",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "twilio>=9.10.4",
]
